pyproject.toml
[tool.poetry]
name = "r2k"
version = "0.7.4"
description = "A tool that lets you periodically send articles received from an RSS feed to your Kindle"
authors = ["Pavel Brodsky <mcouthon@gmail.com>", "Auguste Baum <auguste.baum@protonmail.com>"]
license = "MIT"
repository = "https://github.com/mcouthon/r2k"
homepage = "https://github.com/mcouthon/r2k"
keywords = ["rss", "atom", "kindle", "reading"]
readme = "README.md"

packages = [{ include = "r2k" }]

include = ["templates/**/*"]

[tool.poetry.scripts]
r2k = "r2k.cli:main"

[tool.poetry.dependencies]
python = "^3.8"

arrow = "^1.3.0"
beautifulsoup4 = "^4.12.3"
click = "^8.1.7"
feedparser = "^6.0.0"
orjson = "^3.9.14"
pick = "^2.2.0"
pydantic = "^2.6.1"
python-dateutil = "^2.8.2"
pyyaml = "^6.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
# pre-commit = "^2.2.0"
# isort = "^4.3.21"
# black = "^21.10b0"
# pyupgrade = "^2.2.0"
# mypy = "^0.770"
# pydocstyle = "^5.0.2"
# pytest = "^5.4.1"

# flake8 = "^4.0.1"
# pycodestyle = "^2.8.0"

[tool.poetry.group.readability.dependencies]
readability-lxml = "^0.8.1"

# Make it optional, as it's only relevant if we're using the mercury parser API
# docker = { version = "^4.2.0", optional = true }

[tool.poetry.extras]
docker = ["docker"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry>=1.1.3"]
build-backend = "poetry.core.masonry.api"
